requirements.txt
telethon
tweepy
pysocks
configparser